Reorganize address handler unit tests into separate test files for each individual class
